+++ src/Tests/Aardvark.Rendering.Tests/Program.fs
@@ -23,6 +23,7 @@ let main argv =
             ``IndexedGeometry Tests``.tests
             ``SceneGraph Tests``.tests
             ``CompactSet Tests``.tests
+            ``ColoredLock Tests``.tests
             ``AdaptiveResource Tests``.tests
             ``ContextCreation Tests``.tests
             ``Vulkan Wrapper Tests``.tests